# docs/public/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Casebox.Server | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/patterns": {
      "get": {
        "tags": [
          "Patterns"
        ],
        "parameters": [
          {
            "name": "workspace",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/patterns/{id}": {
      "get": {
        "tags": [
          "Patterns"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/patterns/{id}/acknowledgement": {
      "post": {
        "tags": [
          "Patterns"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/patterns/{id}/dismissal": {
      "post": {
        "tags": [
          "Patterns"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DismissBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/worker/v1/jobs/lease": {
      "post": {
        "tags": [
          "Worker"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worker/v1/jobs/{id}/heartbeat": {
      "post": {
        "tags": [
          "Worker"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerRef"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worker/v1/jobs/{id}/complete": {
      "post": {
        "tags": [
          "Worker"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/worker/v1/jobs/{id}/fail": {
      "post": {
        "tags": [
          "Worker"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompleteRequest": {
        "required": [
          "workerId",
          "result"
        ],
        "type": "object",
        "properties": {
          "workerId": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/JsonElement"
          }
        }
      },
...
      "DismissBody": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "FailRequest": {
        "required": [
          "workerId",
          "error",
          "retryable"
        ],
        "type": "object",
        "properties": {
          "workerId": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
...
      "JsonElement": {},
      "LeaseRequest": {
        "required": [
          "workerId",
          "version",
          "kinds"
        ],
        "type": "object",
        "properties": {
          "workerId": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "kinds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "WorkerRef": {
        "required": [
          "workerId"
        ],
        "type": "object",
        "properties": {
          "workerId": {
            "type": "string"
          }
        }
      }
    }
  },
...
}